Fixed grass creation function issue
Logic problem, remember functions need to return something otherwise
they return void. When you assign a function to a variable, even if it
changes some global variable (e.g. array value somewhere outside it's
scope), it will return void to the variable assigned to it (in case it
is the same element of the array for example that you manipulated with
the function, if you assign that function to that array position it will
void it after creating/modifying it).